NASAEARTH-16 add time core graph to the plugin
This is a copy of the graph created in the dst standalone application: https://github.com/concord-consortium/dst

It uses chart.js to render the graph. Each color is a separate chart.js dataset. And then the dataset defines the "points" which are pin labels on the x-axis and date ranges on the y-axis

The css is adjusted so the graph can take the maximum height when the plugin is resized in CODAP.

Changes to the DataManager:
- it now has a singleton instance which is used by both the ColorbarPlot and the DatasetTab.
- it tracks the currentImageIndex
- it is observable so the ColorbarPlot can update when the images, pins, or currentImageIndex changes
- the chart is no longer created because that is now replaced by the internal ColorbarPlot

The initializePlugin is awaited now, this gives CODAP a little more time to adjust the size of the plugin before the map is added. This doesn't always work, but it sometimes help the map show up next to the plugin better.

Things that are not good:
- the tooltips haven't been customized so they are hard to read. They also cover the graph.
- if the dataset is missing a month this gap is not shown on the plot

Things that should probably not be implemented:
- dragging the annotation: this would not be supported if the graph is moved into CODAP itself, and would also be harder to implement if the graph was turned into a separate plugin so it could be laid out better.